refactor: Restructure package hierarchy for UI components
